# LibGfx: Make webp lossless decoder 6 times as fast